Move class assignments to dedicated subpage
